requirements.txt
attrs==21.4.0
bidict==0.22.0
black==22.3.0
Flask==2.1.2
Flask-Cors==3.0.10
flask-smorest==0.37.0
Flask-SQLAlchemy==2.5.1
marshmallow==3.15.0
pytest==7.1.2
python-dotenv==0.20.0
python-jwt==2.0.1
requests
SQLAlchemy==1.4.36
tomli==2.0.1
coloredlogs==15.0.1
psycopg2-binary
gevent
pyjwt
pandas
boto3
snowflake-sqlalchemy